refactored giant slog of code in the main index into routes, needs further refactoring to modules to get it to be nicer
