+++ .eslintrc.json
@@ -18,6 +18,8 @@
     "object-curly-spacing": [ "error", "always" ],
     "semi-spacing": "error",
     "no-confusing-arrow": "error",
-    "no-constant-condition": "error"
+    "no-constant-condition": "error",
+    "keyword-spacing": "error",
+    "no-unused-vars": "error"
   }
 }
