+++ api/routes.js
@@ -1,0 +1,154 @@
+const { Bleetify } = require('./bleetify');
+const { handleRoles } = require('./roles');
+
+const Dice = require('./dice');
+const Pjson = require('../package.json');
+const Util = require('./util');
+const Vote = require('./vote');
+
+const routes = {
+  badgoat: {
+    name: 'badgoat',
+    description: 'downvote the goat :(',
+    command: '!badgoat',
+    example: '!badgoat',
+    execute: Vote.badGoat
+  },
+  balance: {
+    name: 'balance',
+    description: 'the goats status, how good or bad it\'s has been, synonymous with !count',
+    command: '!balance',
+    example: '!balance or !count',
+    execute: Vote.balance
+  },
+  colour: {
+    name: 'colour',
+    description: 'change your name colour',
+    command: '!colour',
+    example: '!colour #97dbc8',
+    execute: (message) => {
+      const hex = message.content.split(' ')[1];
+      const role = message.member.highestRole;
+      role.setColor(hex)
+        .then(() => {
+          message.channel.send(Bleetify(`Your colour is now ${hex}`));
+        })
+        .catch(console.error);
+    }
+  },
+  commands: {
+    name: 'commands',
+    description: 'lists the available commands synonymous with !help without ',
+    command: '!commands',
+    example: '!commands or !help',
+    execute: (message) => help(message)
+  },
+  count: {
+    name: 'count',
+    description: 'the goats status, how good or bad it\'s has been, synonymous with !balance',
+    command: '!count',
+    example: '!count or !balance',
+    execute: Vote.balance
+  },
+  goodgoat: {
+    name: 'goodgoat',
+    description: 'upvote the goat! :3',
+    command: '!goodgoat',
+    example: '!goodgoat',
+    execute: Vote.badGoat
+  },
+  hangman: {
+    name: 'hangman',
+    description: 'start a game of hangman or find out about the status of a game',
+    command: '!hangman',
+    example: '!hangman or a letter like !a !p !u'
+  },
+  help: {
+    name: 'help',
+    description: 'lists all available commands or if you specify a command the help for that command',
+    command: '!help',
+    example: '!help or !commands',
+    execute: (message) => help(message)
+  },
+  ping: {
+    name: 'ping',
+    description: 'responds with pong',
+    command: '!ping',
+    example: '!ping',
+    execute: () => message.channel.send(Bleetify('Pong!'))
+  },
+  pong: {
+    name: 'pong',
+    description: 'responds with ping',
+    command: '!pong',
+    example: '!pong',
+    execute: () => message.channel.send(Bleetify('Ping!'))
+  },
+  role: {
+    name: 'role',
+    description: 'Set your role to one of the available non-admin roles',
+    command: '!role',
+    example: '!role overwatch',
+    execute: handleRoles
+  },
+  roll: {
+    name: 'roll',
+    description: 'roll a number of dice, with specific size, targets to get above or below and modifiers you can add',
+    command: '!roll',
+    example: '!roll 2d10>8+2 you\'re rolling 2 dice that have 10 sides, aiming for 8 or higher and +2 to each of your rolls',
+    execute: (message) => {
+      const { numSides, diceToRoll, modifier, challenge, command } = Dice.getDiceOptions(content);
+      const result = Dice.roll(numSides, diceToRoll, modifier, challenge);
+      return message.channel.send(Bleetify(`You rolled ${command} and got; ${result}`));
+    }
+  },
+  sexygoat: {
+    name: 'sexygoat',
+    description: 'people being weird to th egoat is not cool man',
+    command: '!sexygoat',
+    example: '!sexygoat - dont',
+    execute: () => message.channel.send("``` You've really freaked the goat out :/ don't do that ```")
+  },
+  upcoming: {
+    name: 'upcoming',
+    description: 'what we are currently working on, features to come',
+    command: '!upcoming',
+    example: '!upcoming - Building some sandcastles :o',
+    execute: () => message.channel.send(Bleetify(workingon))
+  },
+  uptime: {
+    name: 'uptime',
+    description: 'how long the server has been up, how long the goat has been awake',
+    command: '!uptime',
+    example: '!uptime - Baa, we\'ve been up for 2 days',
+    execute: (message) => message.channel.send(Bleetify(`We've been up for ${Util.msToTime(client.uptime)}`))
+  },
+  version: {
+    name: 'version',
+    description: 'what version we are running, this is important for debugging',
+    command: '!version',
+    example: '!version - v3.0.2',
+    execute: () => message.channel.send(Bleetify(`We're using v${Pjson.version}`))
+  },
+  workingon: {
+    name: 'working on',
+    description: 'what we are currently working on, upcoming features or fixes',
+    command: '!workingon',
+    example: '!workingon - We\'re currently working on: maintainability! This grew really fast and needs fixing!',
+    execute: () => message.channel.send(Bleetify(workingon))
+  }
+};
+
+const workingon = `I'm working on maintainability at the moment to help with the ease of adding features into the future`;
+
+const help = function (message) {
+  const commandsList = [];
+  Object.keys(routes).forEach(key => {
+    commandsList.push(`  * ${key.example}`);
+  });
+  return message.channel.send(`\`\`\` Available commands are:
+  ${commandsList.toString('\n')}
+  \`\`\``);
+};
+
+module.exports = routes;
